feat: restructure playground files
