--- playground/index.css
@@ -1,4 +1,0 @@
-@import 'tailwindcss';
-@plugin "daisyui" {
-  themes: light --default;
-}
