+++ playground/index.html
@@ -7,7 +7,7 @@
     <title>Playground</title>
   </head>
   <body>
-    <div id="root"></div>
-    <script type="module" src="/index.tsx"></script>
+    <input type="file" />
+    <script type="module" src="/main.ts"></script>
   </body>
 </html>
